feat: dynamic jump analysis
