+++ crates/dora-compiler/src/evm/program.rs
@@ -1,4 +1,3 @@
-use dora_primitives::spec::SpecId;
 use num_bigint::BigUint;
 pub use revmc::{op_info_map, OpcodeInfo};
 use std::fmt;
@@ -417,31 +416,6 @@ operations!(
     (SelfDestruct, SELFDESTRUCT),
 );
 
-/// Represents the options used during the compilation process.
-/// This struct encapsulates various settings that can be adjusted to customize the compilation behavior.
-#[derive(Debug)]
-pub struct CompileOptions {
-    /// Specification IDs and their activation block.
-    ///
-    /// Information was obtained from the [Ethereum Execution Specifications](https://github.com/ethereum/execution-specs)
-    pub spec_id: SpecId,
-    /// A flag indicating whether to perform gas metering during compilation.
-    pub gas_metering: bool,
-    /// Check for stack overflow or underflow errors. Note that there is no need to check for EOF Bytecode,
-    /// as stack operations are statically determined at compile time.
-    pub stack_bound_checks: bool,
-}
-
-impl Default for CompileOptions {
-    fn default() -> Self {
-        Self {
-            spec_id: SpecId::CANCUN,
-            gas_metering: true,
-            stack_bound_checks: true,
-        }
-    }
-}
-
 /// Represents a program that has been parsed and is ready for execution. The `Program` struct
 /// holds a list of operations and the total code size of the bytecode it represents.
 ///
@@ -539,6 +513,30 @@ impl Program {
             .iter()
             .flat_map(Operation::to)
             .collect::<Vec<u8>>()
+    }
+
+    /// Returns `true` if the EVM program is EOF.
+    #[inline]
+    pub fn is_eof(&self) -> bool {
+        false
+    }
+
+    /// Mark `PUSH<N>` followed by `JUMP[I]` as `STATIC_JUMP` and resolve the target.
+    pub fn has_dynamic_jumps(&mut self) -> bool {
+        debug_assert!(!self.is_eof());
+        for i in 0..self.operations.len() {
+            let op = &self.operations[i];
+            let is_jump = matches!(op, Operation::Jump | Operation::Jumpi);
+            let is_push = i > 0
+                && matches!(
+                    self.operations[i - 1],
+                    Operation::Push0 | Operation::Push(_)
+                );
+            if is_jump && !is_push {
+                return true;
+            }
+        }
+        false
     }
 
     fn parse_operations(opcodes: &[u8]) -> (Vec<Operation>, Vec<OpcodeParseError>) {
